requirements.txt
streamlit>=1.30.0
google-genai>=0.1.0
google-generativeai>=0.5.0
pillow>=10.0.0
pytest>=8.0.0
google-cloud-logging>=3.0.0
pytest-mock>=3.10.0
